mgr/dashboard_v2: use float values when checking session expire time

+++ src/pybind/mgr/dashboard_v2/auth.py
@@ -25,7 +25,7 @@ class Auth(object):
     SESSION_KEY = '_username'
     SESSION_KEY_TS = '_username_ts'
 
-    DEFAULT_SESSION_EXPIRE = 1200
+    DEFAULT_SESSION_EXPIRE = 1200.0
 
     @staticmethod
     def password_hash(password, salt_password=None):
@@ -44,7 +44,7 @@ class Auth(object):
     @cherrypy.tools.allow(methods=['POST'])
     @tools.json_out()
     def login(self, username=None, password=None):
-        now = int(time.time())
+        now = time.time()
         config_username = self.module.get_localized_config('username', None)
         config_password = self.module.get_localized_config('password', None)
         hash_password = Auth.password_hash(password,
@@ -74,13 +74,13 @@ class Auth(object):
                 relative='server')))
             raise cherrypy.HTTPError(401, 'You are not authorized to access '
                                           'that resource')
-        now = int(time.time())
-        expires = int(self.module.get_localized_config(
+        now = time.time()
+        expires = float(self.module.get_localized_config(
                         'session-expire',
                         Auth.DEFAULT_SESSION_EXPIRE))
         if expires > 0:
             username_ts = cherrypy.session.get(Auth.SESSION_KEY_TS, None)
-            if username_ts and username_ts < now - expires:
+            if username_ts and float(username_ts) < (now - expires):
                 cherrypy.session[Auth.SESSION_KEY] = None
                 cherrypy.session[Auth.SESSION_KEY_TS] = None
                 self.log.debug('Session expired.')
